requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
python-gitlab==2.9.0
requests==2.25.1
requests-toolbelt==0.9.1
sty==1.0.0rc1
urllib3==1.26.6
